feat(66-02): enable -Werror=switch-enum globally and audit every Iron-enum switch
PROT-02: Append -Werror=switch-enum to the non-MSVC add_compile_options line
in CMakeLists.txt so iron_compiler, iron_runtime, iron_stdlib, and ironc all
build under the new strictness posture. Adding a new variant to any Iron enum
without updating every consumer now becomes a compile error, structurally
preventing the silently-missing-case bug class that motivated Phase 66 (rank
13 collect_mono_enums_node walkthrough).

- CMakeLists.txt: append -Werror=switch-enum to the shared -Wall -Wextra
  -Werror -Wpedantic line (applies globally to all non-MSVC targets;
  Unity carve-out in the MSVC branch already isolates the third-party lib)

- AUDIT-02 #1 src/parser/printer.c op_str: add cases for every bitwise
  operator (AMP/PIPE/CARET/TILDE/SHL/SHR) and compound-assign siblings;
  opt-out the remaining non-operator tokens

- AUDIT-02 #2-4 src/hir/hir_lower.c: enumerate every legitimate operator in
  ast_op_to_hir_binop, compound_assign_base_op, and the unary-op mapping;
  defensive fallbacks are retained with opt-out comments

- AUDIT-02 #5 src/hir/hir_to_lir.c collect_mono_enums_node: add explicit
  recursion for FOR, DEFER, SPAWN, FREE, LEAK, LAMBDA, INDEX, SLICE,
  ARRAY_LIT, CONSTRUCT, HEAP, RC, IS, AWAIT, COMPTIME, INTERP_STRING — the
  8+ kinds that previously fell through and caused the SIGSEGV walkthrough

- AUDIT-02 #6 src/hir/hir_to_lir.c elem_suffix: add INT8/INT16/INT64,
  UINT family, FLOAT32/FLOAT64; previously narrow/wide int and float kinds
  silently dispatched to Iron_List_int64_t_*

- AUDIT-02 #7 src/hir/hir_verify.c verify_expr: add ENUM_CONSTRUCT and
  PATTERN arms recursing into sub-expressions / nested patterns

- AUDIT-02 #8 src/comptime/comptime.c eval_expr: enumerate METHOD_CALL,
  FIELD_ACCESS, INDEX, SLICE, INTERP_STRING, LAMBDA, IS, MATCH as explicit
  'unsupported in comptime' diagnostics

- AUDIT-02 #9 src/comptime/comptime.c val_to_ast + cache read/write:
  enumerate STRUCT, NULL, ARRAY explicitly so the exhaustive CVAL_ switch
  is literal and the default is removed from the terminal switch

- AUDIT-02 #10-11 src/analyzer/escape.c collect_stmt + validate_node: add
  MATCH, DEFER, SPAWN recursion so heap bindings / free / leak inside
  these nested scopes are no longer silently evaded

- AUDIT-02 #12 src/analyzer/concurrency.c collect_spawn_refs: add CALL,
  METHOD_CALL, MATCH, DEFER, FREE, LEAK recursion

- AUDIT-02 #15-16 src/lir/lir_optimize.c analyze_array_param_modes +
  collect_mono_enums_node copy: opt-out the array-param scanner (inspects
  only a subset of IronLIR_OpKind) and add the same 16 missing kinds to
  the optimizer's copy of collect_mono_enums_node

- src/lir/emit_c.c FFI cast logic (~L2820): rewrite the if-chain on
  arg_type->kind as an exhaustive switch over Iron_TypeKind with a
  commented default, preserving byte-for-byte semantics (STRING /
  INT / INT64 / FLOAT64 are the only kinds the FFI bridge rewrites)

- src/lir/lir_optimize.c, src/lir/value_range.c, src/lir/verify.c,
  src/lir/lir.c, src/lir/print.c: opt-out the LIR walkers that consume
  only a subset of the 44 IronLIR_OpKind variants (escape scan, dead
  alloca, store-load elim, constant folding, operand collectors,
  memory-mutation predicate, purity predicate, CFG rebuild)

- src/analyzer/typecheck.c, types.c, capture.c, init_check.c, resolve.c:
  opt-out every generic AST walker / type predicate that legitimately
  handles only a subset of Iron_NodeKind / Iron_TypeKind / Iron_SymbolKind

- src/analyzer/web_await_check.c, web_top_level_loader_check.c: opt-out
  the net-new WebAssembly analyzer BFS walkers (merged in PR #23)

- src/parser/parser.c: opt-out iron_infix_prec, iron_parse_primary,
  iron_parse_stmt, iron_parse_decl — each is a Pratt-parser dispatch
  over ~80 token kinds where only operator/keyword prefixes are legal

- src/diagnostics/diagnostics.c iron_diag_emit: convert default: to an
  explicit IRON_DIAG_NOTE case so the 3-variant Iron_DiagLevel enum is
  literally exhaustive

- src/lir/print.c print_instr: convert the unreachable default to
  IRON_LIR_INSTR_COUNT sentinel case

- tests/unit/test_resolver.c find_ident: opt-out the test helper walker

Pattern used for opt-outs: switch ((int)expr->kind) { ... default: ... }
converts the formally-enum switch to a formally-int switch so the flag
no longer trips, while an inline /* -Wswitch-enum opt-out: <reason> */
comment leaves a grep-visible audit trail for future tooling (82 opt-outs
across 21 source files).

Clean cmake build succeeds with zero switch-enum errors or warnings.
All 72 ctest targets pass (unit + hir + lir + algorithms + composite +
integration + tsan). tests/integration/run_integration.sh passes 346/346
with 0 failures. Behavior is unchanged for every test fixture — the fixes
add recursion into kinds that were silently dropped, but none of the
existing fixtures exercised those paths in a way that would change output.

Satisfies PROT-02 and delivers AUDIT-02 M-severity fixes #1-12, #14-16.
AUDIT-02 #13 (typecheck.c covered[256]) is not a switch-enum issue and
is deferred to Phase 67 per plan. #17 (iron_net.c errno/WSA translation)
switches are over int, not Iron enums, and are unaffected by the new
flag — no changes needed.

+++ src/analyzer/capture.c
@@ -42,7 +42,7 @@ typedef struct {
 
 /* Returns true if a symbol kind should never be treated as a capture. */
 static bool sym_kind_is_non_capture(Iron_SymbolKind k) {
-    switch (k) {
+    switch ((int)(k)) {
         case IRON_SYM_FUNCTION:
         case IRON_SYM_METHOD:
         case IRON_SYM_TYPE:
@@ -51,6 +51,8 @@ static bool sym_kind_is_non_capture(Iron_SymbolKind k) {
         case IRON_SYM_INTERFACE:
         case IRON_SYM_FIELD:
             return true;
+        /* -Wswitch-enum opt-out: IRON_SYM_VARIABLE and IRON_SYM_PARAM ARE
+         * capturable and intentionally fall through to `false`. */
         default:
             return false;
     }
@@ -63,7 +65,7 @@ static bool sym_kind_is_non_capture(Iron_SymbolKind k) {
  * val/var decls inside the lambda body. */
 static void collect_locals(Iron_Node *node, StrSet **locals) {
     if (!node) return;
-    switch (node->kind) {
+    switch ((int)(node->kind)) {
         case IRON_NODE_VAL_DECL: {
             Iron_ValDecl *vd = (Iron_ValDecl *)node;
             shput(*locals, vd->name, 1);
@@ -119,6 +121,9 @@ static void collect_locals(Iron_Node *node, StrSet **locals) {
         }
         /* Do NOT recurse into nested IRON_NODE_LAMBDA — its locals are
          * handled when find_captures processes it separately. */
+        /* -Wswitch-enum opt-out: collect_locals only visits statement-shaped
+         * kinds that can declare a new name; every other Iron_NodeKind is
+         * ignored. */
         default:
             break;
     }
@@ -135,7 +140,7 @@ static void walk_node_for_lambdas(CaptureCtx *ctx, Iron_Node *node);
 static void collect_idents(Iron_Node *node, StrSet **locals,
                             TmpCapture **captures, StrSet **seen) {
     if (!node) return;
-    switch (node->kind) {
+    switch ((int)(node->kind)) {
         case IRON_NODE_IDENT: {
             Iron_Ident *id = (Iron_Ident *)node;
             if (!id->resolved_sym) break;
@@ -347,6 +352,9 @@ static void collect_idents(Iron_Node *node, StrSet **locals,
             collect_idents(ss->body,      locals, captures, seen);
             break;
         }
+        /* -Wswitch-enum opt-out: collect_idents walks every expression-bearing
+         * kind; leaf nodes (literals, self, super) have no identifiers of
+         * their own and intentionally fall through. */
         default:
             break;
     }
@@ -478,7 +486,7 @@ static void find_pfor_captures(CaptureCtx *ctx, Iron_ForStmt *fs) {
  * ordering), then process the lambda itself. */
 static void walk_node_for_lambdas(CaptureCtx *ctx, Iron_Node *node) {
     if (!node) return;
-    switch (node->kind) {
+    switch ((int)(node->kind)) {
         case IRON_NODE_LAMBDA: {
             Iron_LambdaExpr *le = (Iron_LambdaExpr *)node;
             /* Process nested lambdas first */
@@ -670,6 +678,9 @@ static void walk_node_for_lambdas(CaptureCtx *ctx, Iron_Node *node) {
             find_spawn_captures(ctx, ss);
             break;
         }
+        /* -Wswitch-enum opt-out: walk_node_for_lambdas is a generic AST
+         * walker; every kind that does not contain a lambda / spawn / pfor
+         * is a valid no-op. */
         default:
             break;
     }
@@ -681,7 +692,7 @@ static void walk_node_for_lambdas(CaptureCtx *ctx, Iron_Node *node) {
  * Called by iron_capture_verbose_report only when --verbose is active. */
 static void verbose_walk(Iron_Node *node, Iron_Arena *arena, int depth) {
     if (!node) return;
-    switch (node->kind) {
+    switch ((int)(node->kind)) {
         case IRON_NODE_LAMBDA: {
             Iron_LambdaExpr *le = (Iron_LambdaExpr *)node;
             fprintf(stderr, "  [capture] lambda at %s:%u captures %d variable(s):\n",
@@ -788,6 +799,8 @@ static void verbose_walk(Iron_Node *node, Iron_Arena *arena, int depth) {
             }
             break;
         }
+        /* -Wswitch-enum opt-out: verbose_walk is a debugging dumper that
+         * only descends into the kinds it explicitly handles. */
         default:
             break;
     }
@@ -807,7 +820,7 @@ void iron_capture_analyze(Iron_Program *program, Iron_Scope *global_scope,
     for (int i = 0; i < program->decl_count; i++) {
         Iron_Node *decl = program->decls[i];
         if (!decl) continue;
-        switch (decl->kind) {
+        switch ((int)(decl->kind)) {
             case IRON_NODE_FUNC_DECL: {
                 Iron_FuncDecl *fd = (Iron_FuncDecl *)decl;
                 if (fd->body) walk_node_for_lambdas(&ctx, fd->body);
@@ -818,6 +831,8 @@ void iron_capture_analyze(Iron_Program *program, Iron_Scope *global_scope,
                 if (md->body) walk_node_for_lambdas(&ctx, md->body);
                 break;
             }
+            /* -Wswitch-enum opt-out: capture analysis only runs on function /
+             * method bodies. */
             default:
                 break;
         }
@@ -834,7 +849,7 @@ void iron_capture_verbose_report(Iron_Program *program, Iron_Arena *arena) {
     for (int i = 0; i < program->decl_count; i++) {
         Iron_Node *decl = program->decls[i];
         if (!decl) continue;
-        switch (decl->kind) {
+        switch ((int)(decl->kind)) {
             case IRON_NODE_FUNC_DECL: {
                 Iron_FuncDecl *fd = (Iron_FuncDecl *)decl;
                 if (fd->body) verbose_walk(fd->body, arena, 0);
@@ -845,6 +860,8 @@ void iron_capture_verbose_report(Iron_Program *program, Iron_Arena *arena) {
                 if (md->body) verbose_walk(md->body, arena, 0);
                 break;
             }
+            /* -Wswitch-enum opt-out: verbose report only visits func / method
+             * bodies; every other top-level decl kind is intentionally ignored. */
             default:
                 break;
         }
